# docs/03_API仕様/openapi.yaml
openapi: 3.1.0
info:
  title: 人事考課システム API
  description: |
    トキワテック人事考課システムのREST API仕様書。

    ## 認証
    Cloudflare Access (JWT) による認証を使用します。
    全てのAPIリクエストには `Authorization: Bearer <token>` ヘッダーが必要です。

    ## 共通仕様
    - 日時はISO 8601形式 (例: `2026-01-24T10:00:00Z`)
    - IDはUUID v4形式
    - ページネーションはoffset/limit方式
  version: 1.0.0
  contact:
    name: トキワテック開発チーム
    email: dev@tokiwa-tech.com
  license:
    name: Proprietary
    identifier: LicenseRef-proprietary
servers:
  - url: https://evaluation.tokiwa-tech.com/api/v1
    description: 本番環境
  - url: https://evaluation-staging.tokiwa-tech.com/api/v1
    description: ステージング環境
  - url: http://localhost:8787/api/v1
    description: ローカル開発環境
...
paths:
...
  /periods:
    get:
      tags: [Periods]
      summary: 評価期間一覧取得
      operationId: listPeriods
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PeriodStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 評価期間一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodListResponse'

    post:
      tags: [Periods]
      summary: 評価期間作成
      description: 管理者のみ実行可能
      operationId: createPeriod
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePeriodRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /periods/{periodId}:
    get:
      tags: [Periods]
      summary: 評価期間詳細取得
      operationId: getPeriod
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PeriodId'
      responses:
        '200':
          description: 評価期間詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /periods/{periodId}/activate:
    put:
      tags: [Periods]
      summary: 評価期間開始
      description: PREPARING → ACTIVE に遷移。ロードマップのスナップショットを作成
      operationId: activatePeriod
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PeriodId'
      responses:
        '200':
          description: 開始成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodResponse'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /roadmap-items:
    get:
      tags: [Roadmap]
      summary: ロードマップ項目一覧
      operationId: listRoadmapItems
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/AxisCode'
        - name: gradeId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ロードマップ項目一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadmapItemListResponse'
...
  /scenarios:
    get:
      tags: [Scenarios]
      summary: 文章題一覧取得
      operationId: listScenarios
      security:
        - BearerAuth: []
      parameters:
        - name: targetGrade
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ScenarioStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: シナリオ一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioListResponse'

    post:
      tags: [Scenarios]
      summary: 文章題作成
      operationId: createScenario
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScenarioRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
  /scenarios/{scenarioId}:
    get:
      tags: [Scenarios]
      summary: 文章題詳細取得
      operationId: getScenario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ScenarioId'
      responses:
        '200':
          description: シナリオ詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioDetailResponse'

    put:
      tags: [Scenarios]
      summary: 文章題更新
      operationId: updateScenario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ScenarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScenarioRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'

    delete:
      tags: [Scenarios]
      summary: 文章題削除
      description: 管理者のみ
      operationId: deleteScenario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ScenarioId'
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /scenarios/{scenarioId}/questions:
    post:
      tags: [Scenarios]
      summary: 設問追加
      description: HR/管理者のみ
      operationId: addScenarioQuestion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ScenarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioQuestionResponse'
  /scenarios/{scenarioId}/review:
    post:
      tags: [Scenarios]
      summary: シナリオレビュー
      description: HR/管理者のみ
      operationId: reviewScenario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ScenarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved:
                  type: boolean
                comment:
                  type: string
      responses:
        '200':
          description: レビュー完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
  /scenarios/{scenarioId}/publish:
    post:
      tags: [Scenarios]
      summary: シナリオ公開
      description: HR/管理者のみ
      operationId: publishScenario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ScenarioId'
      responses:
        '200':
          description: 公開成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
  /scenarios/{scenarioId}/generate-curriculum:
    post:
      tags: [Scenarios, Curriculum]
      summary: カリキュラム自動生成
      operationId: generateCurriculum
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ScenarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateCurriculumRequest'
      responses:
        '202':
          description: 生成開始（非同期）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumGenerationResponse'

  # ============================================
  # カリキュラム API (11 endpoints)
  # ============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cloudflare Access JWT token
  parameters:
    PeriodId:
      name: periodId
      in: path
      required: true
      schema:
        type: string
      description: 評価期間ID
...
    ScenarioId:
      name: scenarioId
      in: path
      required: true
      schema:
        type: string
      description: シナリオID
...
    PeriodStatus:
      name: status
      in: query
      schema:
        type: string
        enum: [PREPARING, ACTIVE, CLOSED]
      description: 評価期間ステータス
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: ページ番号
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 1ページあたりの件数
  responses:
...
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: リソースなし
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 状態競合
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNext:
          type: boolean

    # ============================================
    # 列挙型
    # ============================================
...
    AxisCode:
      type: string
      enum: [SKILL, COMPETENCY, BEHAVIOR]
...
    ScenarioStatus:
      type: string
      enum: [DRAFT, GENERATED, REVIEWED, PUBLISHED, ARCHIVED]
...
    Difficulty:
      type: string
      enum: [BEGINNER, INTERMEDIATE, ADVANCED, EXPERT]

    # ============================================
    # 認証関連
    # ============================================
...
    Period:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        evaluationMonth:
          type: integer
          minimum: 1
          maximum: 12
        status:
          type: string
          enum: [PREPARING, ACTIVE, CLOSED]
    CreatePeriodRequest:
      type: object
      required: [name, startDate, endDate, evaluationMonth]
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        evaluationMonth:
          type: integer
          minimum: 1
          maximum: 12
    PeriodResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Period'
    PeriodListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Period'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PeriodDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Period'
            - type: object
              properties:
                criteria:
                  type: array
                  items:
                    $ref: '#/components/schemas/EvaluationCriteria'
                cycleStats:
                  type: object
                  properties:
                    total:
                      type: integer
                    byStatus:
                      type: object
                      additionalProperties:
                        type: integer
    EvaluationCriteria:
      type: object
      properties:
        axisCode:
          $ref: '#/components/schemas/AxisCode'
        axisName:
          type: string
        weight:
          type: integer
        maxScore:
          type: integer

    # ============================================
    # 評価サイクル
    # ============================================
...
    RoadmapItem:
      type: object
      properties:
        id:
          type: string
        itemNumber:
          type: integer
        category:
          $ref: '#/components/schemas/AxisCode'
        categoryName:
          type: string
        subcategory:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        levels:
          type: array
          items:
            type: object
            properties:
              gradeId:
                type: string
              gradeCode:
                type: string
              requirement:
                type: string
              isRequired:
                type: boolean
    RoadmapItemListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/RoadmapItem'
...
    Scenario:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        targetGrades:
          type: array
          items:
            type: string
        skillCategory:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        status:
          $ref: '#/components/schemas/ScenarioStatus'
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    ScenarioListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Scenario'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ScenarioResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Scenario'
    ScenarioDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Scenario'
            - type: object
              properties:
                questions:
                  type: array
                  items:
                    $ref: '#/components/schemas/ScenarioQuestion'
                curriculum:
                  $ref: '#/components/schemas/Curriculum'
    ScenarioQuestion:
      type: object
      properties:
        id:
          type: string
        questionNumber:
          type: integer
        questionType:
          type: string
          enum: [MULTIPLE_CHOICE, FREE_TEXT, CODE_REVIEW, SCENARIO_RESPONSE]
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
          nullable: true
        points:
          type: integer
    CreateScenarioRequest:
      type: object
      required: [title, description, targetGrades, skillCategory]
      properties:
        title:
          type: string
        description:
          type: string
        targetGrades:
          type: array
          items:
            type: string
        skillCategory:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
    UpdateScenarioRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        targetGrades:
          type: array
          items:
            type: string
        skillCategory:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        status:
          $ref: '#/components/schemas/ScenarioStatus'
    GenerateCurriculumRequest:
      type: object
      required: [targetGradeId]
      properties:
        targetGradeId:
          type: string
        includeNoise:
          type: boolean
          default: true
        noiseType:
          type: string
    CurriculumGenerationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            curriculumId:
              type: string
            title:
              type: string
            chapters:
              type: array
              items:
                type: object
            estimatedHours:
              type: integer
            status:
              type: string
    Curriculum:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        targetGrades:
          type: array
          items:
            type: string
        estimatedHours:
          type: integer
        status:
          type: string
...
    CreateQuestionRequest:
      type: object
      required: [questionNumber, questionType, questionText, points]
      properties:
        questionNumber:
          type: integer
        questionType:
          type: string
          enum: [MULTIPLE_CHOICE, FREE_TEXT, CODE_REVIEW, SCENARIO_RESPONSE]
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        points:
          type: integer
        relatedRoadmapItems:
          type: array
          items:
            type: string
    ScenarioQuestionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ScenarioQuestion'

    # ============================================
    # Curriculum - 完全な管理
    # ============================================
...